pyproject.toml
[project]
name = "eyesight"
version = "0.1.0"
description = "Gemini Live API interaction with audio/video"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.13.0",
    "mss>=10.0.0",
    "pillow>=11.2.1",
    "pyaudio>=0.2.14",
    "opencv-python>=4.8.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
eyesight = "eyesight.cli:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 80
target-version = "py313"
